Expose async I/O on net45/netstandard2.0 too
